requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
ale-py==0.8.1
altair==4.2.2
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
atomicwrites==1.4.1
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backports.zoneinfo==0.2.1
baostock==0.8.8
bce-python-sdk==0.8.79
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
box2d-py==2.3.5
cached-property==1.5.2
cachetools==4.2.4
catalogue==1.0.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
colorlog==6.7.0
crcmod==1.7
cycler==0.11.0
Cython==0.29.33
datasets==2.10.1
debugpy==1.6.7
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.4
dm-tree==0.1.8
docopt==0.6.2
easydict==1.10
entrypoints==0.4
et-xmlfile==1.1.0
etils==0.9.0
face-alignment==1.3.3
fastapi==0.95.0
fastavro==1.4.9
fasteners==0.17.3
fastjsonschema==2.16.3
ffmpy==0.3.0
filelock==3.10.0
Flask==2.2.3
Flask-Babel==2.0.0
flatbuffers==23.3.3
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gast==0.4.0
gensim==3.8.1
gin-config==0.5.0
gitdb==4.0.10
GitPython==3.1.31
glfw==2.5.7
google-api-core==1.34.0
google-api-python-client==1.12.10
google-apitools==0.5.31
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.32.0
google-cloud-bigquery-storage==2.11.0
google-cloud-bigtable==1.7.0
google-cloud-core==1.7.3
google-cloud-datastore==1.15.3
google-cloud-dlp==3.6.0
google-cloud-language==1.3.0
google-cloud-pubsub==1.7.0
google-cloud-recommendations-ai==0.10.3
google-cloud-spanner==1.19.1
google-cloud-videointelligence==1.16.1
google-cloud-vision==1.0.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.1.0
googleapis-common-protos==1.58.0
gradio==3.22.1
grpc-google-iam-v1==0.12.3
grpcio==1.54.0
grpcio-gcp==0.2.2
gym==0.23.0
gym-notices==0.0.8
h11==0.14.0
h5py==3.8.0
hdfs==2.6.0
httpcore==0.16.3
httplib2==0.19.1
httpx==0.23.3
huggingface-hub==0.13.2
idna==3.4
imageio==2.23.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.16.2
ipympl==0.9.3
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-server==1.24.0
jupyter-ydoc==0.2.4
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyter_server_fileid==0.9.0
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.22.1
keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
libcst==0.4.1
lightning-utilities==0.8.0
linkify-it-py==2.0.0
llvmlite==0.39.1
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.5
moviepy==1.0.3
msgpack==1.0.5
msgpack-numpy==0.4.7.1
mujoco==2.2.0
mujoco-py==2.1.2.14
multidict==6.0.4
multiprocess==0.70.12.2
mypy-extensions==1.0.0
nbclassic==0.5.6
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==2.6.3
nodejs==0.1.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.56.4
numpy==1.21.6
oauth2client==4.1.3
oauthlib==3.2.2
objsize==0.6.1
opencv-python==4.2.0.34
openpyxl==3.1.2
opt-einsum==3.3.0
optional-django==0.1.0
orjson==3.6.6
packaging==23.1
paddle-bfloat==0.1.7
paddle2onnx==1.0.6
paddlefsl==1.1.0
paddlehub==2.3.1
paddlenlp==2.5.2
paddlepaddle-gpu==2.4.2.post116
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
plac==1.1.3
pluggy==1.0.0
proglog==0.1.10
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==3.19.6
psutil==5.9.5
py==1.11.0
pyarrow==6.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.6
pydeck==0.8.1b0
pydot==1.4.2
pydub==0.25.1
pyfakewebcam==0.1.0
pygame==2.1.3
Pygments==2.15.1
pymongo==3.8.0
Pympler==1.0.1
PyOpenGL==3.1.6
pyparsing==2.4.2
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqt5-tools==5.15.9.3.3
pyrsistent==0.19.3
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-json-logger==2.0.7
python-multipart==0.0.6
pythonping==1.1.4
pytorch-lightning==1.9.4
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyVirtualDisplay==3.0
PyWavelets==1.3.0
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
rarfile==4.0
regex==2022.10.31
requests==2.29.0
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.3.2
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.2
sentencepiece==0.1.97
seqeval==1.2.2
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.26.1
streamlit==1.21.0
swig==4.1.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-datasets==4.8.2
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.6.0
tensorflow-probability==0.19.0
tensorflow-serving-api==2.7.0
termcolor==2.3.0
terminado==0.17.1
tf-agents==0.16.0
threadpoolctl==3.1.0
thulac==0.2.2
tifffile==2021.11.2
tinycss2==1.2.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchmetrics==0.11.4
torchvision==0.14.1+cu116
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.27.1
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uc-micro-py==1.0.1
uri-template==1.2.0
uritemplate==3.0.0
urllib3==1.26.15
uvicorn==0.21.1
validators==0.20.0
visualdl==2.4.2
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wrapt==1.15.0
xxhash==3.2.0
y-py==0.5.9
yarl==1.8.2
ypy-websocket==0.8.2
zipp==3.15.0
zstandard==0.21.0

dnspython~=2.3.0
scapy~=2.5.0